hannes: please test before commiting!
From: Oleksandr Motsak <motsak@mathematik.uni-kl.de>

git-svn-id: file:///usr/local/Singular/svn/trunk@13844 2c84dea3-7e68-4137-9b89-c4e89433aadc

+++ kernel/longrat.h
@@ -46,7 +46,7 @@ struct snumber
 
 LINLINE BOOLEAN  nlEqual(number a, number b);
 LINLINE number   nlInit(int i, const ring r);
-number nlRInit (int i);
+number nlRInit (long i);
 LINLINE BOOLEAN  nlIsOne(number a);
 LINLINE BOOLEAN  nlIsZero(number za);
 LINLINE number   nlCopy(number a);
